记录一下，这是完全调通的版本。当然还没有聚类。

+++ src/calculation/calculations.java
@@ -364,7 +364,7 @@ public class calculations {
 				H += calcHk(trail1.getPoints(), trail2.getPoints());
 		}
 		int Pset = topTra.size();
-		H /= Pset;
+		H /= (double)Pset;
 		return H;
 	}
